authJS user login user session prisma model todos per user

+++ src/todos/actions/todo-action.ts
@@ -1,5 +1,6 @@
 "use server";
 
+import { getServerSessionUser } from "@/auth";
 import prisma from "@/lib/prisma";
 import { Todo } from "@prisma/client";
 import { revalidatePath } from "next/cache";
@@ -8,13 +9,14 @@ export const toggleTodoAction = async (
   id: string,
   complete: boolean
 ): Promise<Todo> => {
+  const user = await getServerSessionUser();
   const todo = await prisma.todo.findFirst({ where: { id } });
   if (!todo) {
     throw "Ayoooo this todo does not exist xD";
   }
 
   const updatedTodo = await prisma.todo.update({
-    where: { id },
+    where: { id, userId: user?.id },
     data: { complete },
   });
   revalidatePath("/dashboard/server-todos");
@@ -22,13 +24,22 @@ export const toggleTodoAction = async (
 };
 
 export const addTodoAction = async (description: string): Promise<Todo> => {
-  const todo = await prisma.todo.create({ data: { description } });
+  const user = await getServerSessionUser();
+  const todo = await prisma.todo.create(
+    {
+      data: {
+        description,
+        userId: user?.id,
+      }
+    }
+  );
   revalidatePath("/dashboard/server-todos");
   return todo;
 };
 
 export const deleteCompletedTodosAction = async (): Promise<void> => {
-  const todos = await prisma.todo.deleteMany({ where: { complete: true } });
+  const user = await getServerSessionUser();
+  const todos = await prisma.todo.deleteMany({ where: { complete: true, userId: user?.id } });
   if (todos.count === 0) {
     throw "There are no completed todos to delete";
   }
